==> notebooks/05_combine_data.ipynb <==
# Combine data
Filtered survey data, weights, hooks, and transformed variables

import pandas as pd

# load all data sources

survey_data = pd.read_csv('../data/intermediate/01_filtered_renamed_survey_data.csv').set_index('id')
hooks = pd.read_csv('../data/intermediate/03_survey_population_hooks.csv').set_index('id')
transformed_vars = pd.read_csv('../data/intermediate/04_transformed_variables.csv').set_index('id')
weights = pd.read_csv('../data/intermediate/03_survey_weights.csv').set_index('id')

data = survey_data.join(hooks).join(transformed_vars).join(weights)

data.head()

data.to_csv('../data/processed/survey_data_export.csv', index=True)
